+++ src/main/java/yesman/epicfight/api/event/instance/package-info.java
@@ -1,0 +1,6 @@
+@FieldsAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+package yesman.epicfight.api.event.instance;
+
+import net.minecraft.FieldsAreNonnullByDefault;
+import net.minecraft.MethodsReturnNonnullByDefault;