+++ src/main/java/yesman/epicfight/skill/SkillSlots.java
@@ -2,7 +2,6 @@ package yesman.epicfight.skill;
 
 public enum SkillSlots implements SkillSlot {
 	COMBO_ATTACKS(SkillCategories.BASIC_ATTACK),
-	AIR_SLASH(SkillCategories.AIR_ATTACK),
 	DODGE(SkillCategories.DODGE),
 	PASSIVE1(SkillCategories.PASSIVE),
 	PASSIVE2(SkillCategories.PASSIVE),