+++ src/main/java/yesman/epicfight/world/capabilities/EpicFightCapabilities.java
@@ -1,9 +1,5 @@
 package yesman.epicfight.world.capabilities;
 
-import java.util.Optional;
-
-import org.jetbrains.annotations.Nullable;
-
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
@@ -12,13 +8,19 @@ import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.capabilities.ItemCapability;
 import net.neoforged.neoforge.capabilities.RegisterCapabilitiesEvent;
+import org.jetbrains.annotations.Nullable;
+import yesman.epicfight.client.world.capabilites.entitypatch.player.LocalPlayerPatch;
 import yesman.epicfight.main.EpicFightMod;
 import yesman.epicfight.registry.entries.EpicFightAttachmentTypes;
 import yesman.epicfight.world.capabilities.entitypatch.EntityPatch;
+import yesman.epicfight.world.capabilities.entitypatch.player.PlayerPatch;
+import yesman.epicfight.world.capabilities.entitypatch.player.ServerPlayerPatch;
 import yesman.epicfight.world.capabilities.item.CapabilityItem;
 import yesman.epicfight.world.capabilities.provider.AttachmentEntityPatchProvider;
 import yesman.epicfight.world.capabilities.provider.CommonEntityPatchProvider;
 import yesman.epicfight.world.capabilities.provider.CommonItemCapabilityProvider;
+
+import java.util.Optional;
 
 @EventBusSubscriber(modid = EpicFightMod.MODID)
 public class EpicFightCapabilities {
@@ -40,9 +42,6 @@ public class EpicFightCapabilities {
 	
 	/**
 	 * This method should remain as the secondary option, especially when you can't fix local variables inside lambda expression.
-	 * 
-	 * @param stack
-	 * @return
 	 */
 	public static @Nullable CapabilityItem getItemStackCapability(ItemStack stack) {
 		return getItemCapability(stack).orElse(CapabilityItem.EMPTY);
@@ -50,9 +49,6 @@ public class EpicFightCapabilities {
 	
 	/**
 	 * Return an optional for a given item stack
-	 * 
-	 * @param stack
-	 * @return
 	 */
 	public static Optional<CapabilityItem> getItemCapability(ItemStack stack) {
 		return Optional.ofNullable(stack.getCapability(CAPABILITY_ITEM));
@@ -60,10 +56,9 @@ public class EpicFightCapabilities {
 	
 	/**
 	 * This method should remain as the secondary option, especially when you can't fix local variables inside lambda expression.
-	 * 
+	 * <p>
 	 * @param entity An entity object to extract an entity patch
 	 * @param type A class type to cast
-	 * @return
 	 */
 	@SuppressWarnings("unchecked")
 	public static <T extends EntityPatch<?>> T getEntityPatch(Entity entity, Class<T> type) {
@@ -82,10 +77,9 @@ public class EpicFightCapabilities {
 	/**
 	 * Returns entity patch with unparameterized original entity
 	 * This is useful to reduce the amount of code when type-casting for {@link EntityPatch#getOriginal} is unnecessary.
-	 * 
+	 * <p>
 	 * @param entity An entity object to extract an entity patch
 	 * @param type A class type to cast
-	 * @return
 	 */
 	@SuppressWarnings("unchecked")
 	public static <T extends EntityPatch<?>> Optional<T> getUnparameterizedEntityPatch(Entity entity, Class<T> type) {
@@ -105,11 +99,10 @@ public class EpicFightCapabilities {
 	/**
 	 * Returns entity patch with parameterized original entity
 	 * This method is used when you need parameterized return value of {@link EntityPatch#getOriginal}.
-	 * 
+	 * <p>
 	 * @param entity An entity object to extract an entity patch
 	 * @param entitytype An entity type to cast
 	 * @param patchtype A class type to cast
-	 * @return
 	 */
 	@SuppressWarnings("unchecked")
 	public static <E extends Entity, T extends EntityPatch<E>> Optional<T> getParameterizedEntityPatch(Entity entity, Class<E> entitytype, Class<?> patchtype) {
@@ -124,4 +117,53 @@ public class EpicFightCapabilities {
 		
 		return Optional.empty();
 	}
+
+    /**
+     * Returns {@link PlayerPatch} from a player
+     * @param entity A player to extract the entity patch
+     */
+    public static Optional<PlayerPatch<?>> getPlayerPatch(Entity entity) {
+        AttachmentEntityPatchProvider attachmentEntitypatchProvider = entity.getData(EpicFightAttachmentTypes.ENTITY_PATCH);
+        EntityPatch<?> entitypatch = attachmentEntitypatchProvider.getCapability();
+
+        if (entitypatch instanceof PlayerPatch<?> playerpatch) {
+            return Optional.of(playerpatch);
+        }
+
+        return Optional.empty();
+    }
+
+    /**
+     * Returns {@link LocalPlayerPatch} from a local player
+     * Warning: developers must check physical & logical side before calling this method
+     * <p>
+     * @param entity A player to extract the entity patch
+     */
+    public static Optional<LocalPlayerPatch> getLocalPlayerPatch(Entity entity) {
+        AttachmentEntityPatchProvider attachmentEntitypatchProvider = entity.getData(EpicFightAttachmentTypes.ENTITY_PATCH);
+        EntityPatch<?> entitypatch = attachmentEntitypatchProvider.getCapability();
+
+        if (entitypatch instanceof LocalPlayerPatch localplayerpatch) {
+            return Optional.of(localplayerpatch);
+        }
+
+        return Optional.empty();
+    }
+
+    /**
+     * Returns {@link ServerPlayerPatch} from a server player
+     * Warning: developers must check logical side before calling this method
+     * <p>
+     * @param entity A player to extract the entity patch
+     */
+    public static Optional<ServerPlayerPatch> getServerPlayerPatch(Entity entity) {
+        AttachmentEntityPatchProvider attachmentEntitypatchProvider = entity.getData(EpicFightAttachmentTypes.ENTITY_PATCH);
+        EntityPatch<?> entitypatch = attachmentEntitypatchProvider.getCapability();
+
+        if (entitypatch instanceof ServerPlayerPatch serverplayerpatch) {
+            return Optional.of(serverplayerpatch);
+        }
+
+        return Optional.empty();
+    }
 }